relocate joint loss file
